Fix sandboxing flags handling in browsing context creation
Previously, the algorithm for "determine the sandbox flags" would try to
use whether or not the given browsing context was a child browsing
context, or whether it was an iframe element's nested browsing context.
However, when this algorithm was called during browsing context
creation, the appropriate relationships were not set up yet, so these
conditions would always be false.

This splits the algorithm into two, one for during creation which
accepts an embedder argument indicating the destination container, and
one for post-creation.